backend/requirements.txt
Flask==2.0.1
requests==2.26.0
flask-cors==3.0.10
python-dotenv==0.19.2
groq==0.4.2
google-search-results==2.4.2
feedparser==6.0.8
gunicorn==20.1.0
